Refactoring: Added init file to fix the build. Moved architecture code (e.g. a network without its final layer) into deeplearning/[architecture name]. Moved base network to base/estimators/_network.py. Tests pass.
