+++ sktime_dl/base/estimators/__init__.py
@@ -1,0 +1,3 @@
+from ._classifier import BaseDeepClassifier
+from ._network import BaseDeepNetwork
+from ._regressor import BaseDeepRegressor
